Separate allocator and memory functions
